+++ tests/tests_elements/tests_permutation/test_cases.py
@@ -225,6 +225,17 @@ TEST_IS_REGULAR = [
     (Permutation(2, 1), True),
     (Permutation(2, 1, 3), False),
 ]
+TEST_LEHMER_CODE = [
+    (Permutation(1), [0]),
+    (Permutation(2, 1), [1, 0]),
+    (Permutation(2, 1, 3), [1, 0, 0]),
+    (Permutation(1, 2, 3), [0, 0, 0]),
+    (Permutation(1, 2, 3, 4), [0, 0, 0, 0]),
+    (Permutation(2, 1, 3, 4), [1, 0, 0, 0]),
+    (Permutation(4, 3, 2, 1), [3, 2, 1, 0]),
+    (Permutation(4, 1, 3, 2), [3, 0, 1, 0]),
+    (Permutation(4, 1, 3, 2, 7, 6, 5, 8), [3, 0, 1, 0, 2, 1, 0, 0]),
+]
 TEST_LEXICOGRAPHIC_RANK = [
     (Permutation(1), 1),
     (Permutation(1, 2), 1),